+++ src/properties/export.rs
@@ -4,17 +4,15 @@ use crate::properties::types_json::{
 use bevy::ecs::reflect::ReflectBundle;
 use bevy::reflect::{
     ArrayInfo, EnumInfo, NamedField, StructInfo, TupleInfo, TupleStructInfo, TypeInfo,
-    TypeRegistration, TypeRegistry, UnitVariantInfo, UnnamedField, VariantInfo,
+    TypeRegistration, TypeRegistry, UnnamedField, VariantInfo,
 };
 use bevy::utils::hashbrown::HashMap;
 use bevy::{prelude::*, reflect::ReflectRef};
 use std::borrow::Cow;
-use std::collections::hash_map::Entry;
 use thiserror::Error;
-use tiled::PropertyValue;
 
-const UNIT_COLOR: &str = "#00aa00";
 const DEFAULT_COLOR: &str = "#000000";
+const USE_AS_PROPERTY: &[UseAs] = &[UseAs::Property];
 
 type ExportConversionResult = Result<Vec<TypeExport>, ExportConversionError>;
 
@@ -28,29 +26,49 @@ enum ExportConversionError {
     UnsupportedValue(&'static str),
     #[error("set fields are not supported")]
     SetUnsupported,
-    //#[error("type {0} does not reflect Component, Bundle or Resource")]
-    //NotReflectable(&'static str),
+    #[error("a dependency is not supported")]
+    DependencyError,
 }
 
 #[derive(Debug, Default, Clone)]
 pub(crate) struct TypeExportRegistry {
-    types: HashMap<&'static str, TypeExport>,
+    types: HashMap<&'static str, Vec<TypeExport>>,
     id: u32,
 }
 
 impl TypeExportRegistry {
     #[allow(clippy::wrong_self_convention)]
     pub(crate) fn to_vec(self) -> Vec<TypeExport> {
-        let mut out = self.types.into_values().collect::<Vec<_>>();
+        let mut out = self.types.into_values().flatten().collect::<Vec<_>>();
         out.sort_by(|a, b| a.name.cmp(&b.name));
         out
     }
 
     pub(crate) fn from_registry(registry: &TypeRegistry) -> Self {
+        let mut deps = vec![];
         let mut out = Self::default();
         for t in registry.iter() {
-            out.register_from_type_registration(t, registry);
+            if t.data::<ReflectComponent>().is_some()
+                || t.data::<ReflectBundle>().is_some()
+                || t.data::<ReflectResource>().is_some()
+            {
+                let mut new_deps =
+                    out.register_from_type_registration(t, registry, USE_AS_PROPERTY.to_vec());
+                deps.append(&mut new_deps);
+            }
         }
+
+        // We should have a dedicated 'useAs' flags so we cannot add these dependencies
+        // directly as objects properties (only usable nested)
+        for d in deps {
+            if out.types.contains_key(d) {
+                continue;
+            }
+            if let Some(t) = registry.get_with_type_path(d) {
+                out.register_from_type_registration(t, registry, USE_AS_PROPERTY.to_vec());
+            }
+        }
+
         out
     }
 
@@ -63,24 +81,42 @@ impl TypeExportRegistry {
         &mut self,
         registration: &TypeRegistration,
         registry: &TypeRegistry,
-    ) {
-        match self.generate_export(registration, registry) {
+        use_as: Vec<UseAs>,
+    ) -> Vec<&'static str> {
+        let mut deps = vec![];
+        match self.generate_export(registration, registry, use_as, &mut deps) {
             Ok(export) => {
-                for export in export {
+                if !export.is_empty() {
                     self.types
                         .insert(registration.type_info().type_path(), export);
                 }
+                deps
             }
             Err(_) => {
                 self.remove_with_dependency(registration.type_info().type_path());
+                vec![]
             }
         }
+    }
+
+    fn is_supported(registration: &TypeRegistration) -> bool {
+        matches!(
+            registration.type_info(),
+            TypeInfo::TupleStruct(_)
+                | TypeInfo::Struct(_)
+                | TypeInfo::Tuple(_)
+                | TypeInfo::Array(_)
+                | TypeInfo::Enum(_)
+                | TypeInfo::Opaque(_)
+        )
     }
 
     fn generate_export(
         &mut self,
         registration: &TypeRegistration,
         registry: &TypeRegistry,
+        use_as: Vec<UseAs>,
+        deps: &mut Vec<&'static str>,
     ) -> ExportConversionResult {
         let mut default_value = None;
         let tmp;
@@ -90,25 +126,9 @@ impl TypeExportRegistry {
             default_value = Some(tmp.as_ref());
         }
 
-        let use_as;
-        if registration.data::<ReflectResource>().is_some() {
-            use_as = vec![UseAs::Map];
-        } else if registration.data::<ReflectComponent>().is_some()
-            || registration.data::<ReflectBundle>().is_some()
-        {
-            use_as = UseAs::all_supported();
-        } else {
-            // TODO: right now, we need to export all registered types even if we won't be able to use them
-            // We will change that when we have implemented a way to recursively add types
-            use_as = vec![];
-            // return Err(ExportConversionError::NotReflectable(
-            //    registration.type_info().type_path(),
-            // ));
-        }
-
-        match registration.type_info() {
+        let out = match registration.type_info() {
             TypeInfo::TupleStruct(info) => {
-                self.generate_tuple_struct_export(info, registry, default_value)
+                self.generate_tuple_struct_export(info, registry, default_value, use_as)
             }
             TypeInfo::Struct(info) => {
                 self.generate_struct_export(info, registry, default_value, use_as)
@@ -119,29 +139,47 @@ impl TypeExportRegistry {
             TypeInfo::List(_) => Err(ExportConversionError::ListUnsupported),
             TypeInfo::Array(info) => self.generate_array_export(info, registry, use_as),
             TypeInfo::Map(_) => Err(ExportConversionError::MapUnsupported),
-            TypeInfo::Enum(info) => self.generate_enum_export(info, registry),
+            TypeInfo::Enum(info) => self.generate_enum_export(info, registry, use_as),
             TypeInfo::Opaque(_) => Ok(vec![]),
             TypeInfo::Set(_) => Err(ExportConversionError::SetUnsupported),
+        };
+
+        if out.is_ok() {
+            let mut new_deps = dependencies(registration, registry);
+            if new_deps.iter().all(|n| {
+                if let Some(t) = registry.get_with_type_path(n) {
+                    return Self::is_supported(t);
+                }
+                false
+            }) {
+                deps.append(&mut new_deps);
+                return out;
+            } else {
+                return Err(ExportConversionError::DependencyError);
+            }
         }
+        out
     }
 
     fn remove_with_dependency(&mut self, type_path: &str) {
         let mut to_remove = vec![type_path.to_string()];
         while let Some(type_path) = to_remove.pop() {
-            self.types.retain(|_, export| match &export.type_data {
-                TypeData::Enum(_) => true,
-                TypeData::Class(class) => {
-                    if class.members.iter().any(|m| {
-                        m.property_type
-                            .as_ref()
-                            .is_some_and(|s| s.as_str() == type_path)
-                    }) {
-                        to_remove.push(export.name.clone());
-                        false
-                    } else {
-                        true
+            self.types.retain(|_, export| {
+                export.iter().all(|export| match &export.type_data {
+                    TypeData::Enum(_) => true,
+                    TypeData::Class(class) => {
+                        if class.members.iter().any(|m| {
+                            m.property_type
+                                .as_ref()
+                                .is_some_and(|s| s.as_str() == type_path)
+                        }) {
+                            to_remove.push(export.name.clone());
+                            false
+                        } else {
+                            true
+                        }
                     }
-                }
+                })
             })
         }
     }
@@ -151,12 +189,13 @@ impl TypeExportRegistry {
         info: &TupleStructInfo,
         registry: &TypeRegistry,
         default_value: Option<&dyn Reflect>,
+        _use_as: Vec<UseAs>,
     ) -> ExportConversionResult {
         let root = TypeExport {
             id: self.next_id(),
             name: info.type_path().to_string(),
             type_data: TypeData::Class(Class {
-                use_as: UseAs::all_supported(),
+                use_as: USE_AS_PROPERTY.to_vec(),
                 color: DEFAULT_COLOR.to_string(),
                 draw_fill: true,
                 members: info
@@ -286,7 +325,8 @@ impl TypeExportRegistry {
     fn generate_enum_export(
         &mut self,
         info: &EnumInfo,
-        _registry: &TypeRegistry,
+        registry: &TypeRegistry,
+        _use_as: Vec<UseAs>,
     ) -> ExportConversionResult {
         let simple = info.iter().all(|s| matches!(s, VariantInfo::Unit(_)));
 
@@ -301,7 +341,130 @@ impl TypeExportRegistry {
                 }),
             }])
         } else {
-            Err(ExportConversionError::UnsupportedValue(info.type_path()))
+            // Creates types for:
+            // Enum for the enum variant
+            // Class's for each non-unit variant
+            // Class to hold the variant + each non-unit variant.
+
+            // Note: extra `:` is done to not conflict with an enum variant named Variant
+            let variants_name = info.type_path().to_string() + ":::Variant";
+
+            let mut out = vec![TypeExport {
+                id: self.next_id(),
+                name: variants_name.clone(),
+                type_data: TypeData::Enum(Enum {
+                    storage_type: StorageType::String,
+                    values_as_flags: false,
+                    values: info.iter().map(|s| s.name().to_string()).collect(),
+                }),
+            }];
+
+            let mut root_members = Vec::with_capacity(2);
+            root_members.push(Member {
+                // `:` is to separate from an enum variant named `variant`
+                // and put it at the top of the fields (they are alphabetized in the editor)
+                name: ":variant".to_string(),
+                property_type: Some(variants_name),
+                type_field: FieldType::Class,
+                value: info
+                    .iter()
+                    .next()
+                    .map(|s| serde_json::Value::String(s.name().to_string()))
+                    .unwrap_or_default(),
+            });
+
+            for variant in info.iter() {
+                match variant {
+                    VariantInfo::Struct(s) => {
+                        let name = format!("{}::{}", info.type_path(), s.name());
+                        let import = TypeExport {
+                            id: self.next_id(),
+                            name: name.clone(),
+                            type_data: TypeData::Class(Class {
+                                use_as: USE_AS_PROPERTY.to_vec(),
+                                color: DEFAULT_COLOR.to_string(),
+                                draw_fill: true,
+                                members: s
+                                    .iter()
+                                    .map(|s| {
+                                        let (type_field, property_type) =
+                                            type_to_field(registry.get(s.type_id()).unwrap())?;
+
+                                        Ok(Member {
+                                            name: s.name().to_string(),
+                                            property_type,
+                                            type_field,
+                                            value: Default::default(),
+                                        })
+                                    })
+                                    .collect::<Result<_, _>>()?,
+                            }),
+                        };
+                        out.push(import);
+
+                        let root_field = Member {
+                            name: s.name().to_string(),
+                            property_type: Some(name),
+                            type_field: FieldType::Class,
+                            value: Default::default(),
+                        };
+
+                        root_members.push(root_field);
+                    }
+                    VariantInfo::Tuple(tuple) => {
+                        let name = format!("{}::{}", info.type_path(), tuple.name());
+                        let import = TypeExport {
+                            id: self.next_id(),
+                            name: name.clone(),
+                            type_data: TypeData::Class(Class {
+                                use_as: USE_AS_PROPERTY.to_vec(),
+                                color: DEFAULT_COLOR.to_string(),
+                                draw_fill: true,
+                                members: tuple
+                                    .iter()
+                                    .map(|s| {
+                                        let (type_field, property_type) =
+                                            type_to_field(registry.get(s.type_id()).unwrap())?;
+
+                                        Ok(Member {
+                                            name: s.index().to_string(),
+                                            property_type,
+                                            type_field,
+                                            value: Default::default(),
+                                        })
+                                    })
+                                    .collect::<Result<_, _>>()?,
+                            }),
+                        };
+                        out.push(import);
+
+                        let root_field = Member {
+                            name: tuple.name().to_string(),
+                            property_type: Some(name),
+                            type_field: FieldType::Class,
+                            value: Default::default(),
+                        };
+
+                        root_members.push(root_field);
+                    }
+                    VariantInfo::Unit(_) => continue,
+                }
+            }
+
+            let root = TypeExport {
+                id: self.next_id(),
+                name: info.type_path().to_string(),
+                type_data: TypeData::Class(Class {
+                    use_as: USE_AS_PROPERTY.to_vec(),
+                    color: DEFAULT_COLOR.to_string(),
+                    draw_fill: true,
+                    members: root_members,
+                }),
+            };
+
+            out.push(root);
+
+            Ok(out)
         }
     }
 }
@@ -367,7 +530,6 @@ fn value_to_json(value: &dyn PartialReflect) -> serde_json::Value {
             if info.iter().all(|v| matches!(v, VariantInfo::Unit(_))) {
                 serde_json::json!(v.variant_name())
             } else {
-                // TODO: non-unit enums
                 serde_json::Value::default()
             }
         }
@@ -394,20 +556,20 @@ fn value_to_json(value: &dyn PartialReflect) -> serde_json::Value {
             })
             .collect(),
         _ => {
-            warn!(
-                "cannot convert type '{}' to a JSON value",
-                type_info.type_path()
-            );
+            // warn!(
+            //     "cannot convert type '{}' to a JSON value",
+            //     type_info.type_path()
+            // );
             serde_json::Value::default()
         }
     }
 }
 
 fn named_field_json_value(
-    value: Option<&dyn PartialReflect>,
+    parent_value: Option<&dyn PartialReflect>,
     field: &NamedField,
 ) -> serde_json::Value {
-    match value {
+    match parent_value {
         Some(v) => match v.reflect_ref() {
             ReflectRef::Struct(t) => t
                 .field(field.name())
@@ -420,10 +582,10 @@ fn named_field_json_value(
 }
 
 fn unnamed_field_json_value(
-    value: Option<&dyn PartialReflect>,
+    parent_value: Option<&dyn PartialReflect>,
     field: &UnnamedField,
 ) -> serde_json::Value {
-    match value {
+    match parent_value {
         Some(v) => match v.reflect_ref() {
             ReflectRef::TupleStruct(t) => (*t)
                 .field(field.index())
@@ -489,183 +651,43 @@ fn is_enum_and_simple(t: &TypeRegistration) -> bool {
     }
 }
 
-#[allow(dead_code)]
-fn unit_variant_to_export(info: &UnitVariantInfo, id: u32) -> TypeExport {
-    unit_export(id, info.name().to_string())
-}
+fn dependencies(registration: &TypeRegistration, registry: &TypeRegistry) -> Vec<&'static str> {
+    let deps = match registration.type_info() {
+        TypeInfo::Struct(info) => info.iter().map(NamedField::type_path).collect(),
+        TypeInfo::TupleStruct(info) => info.iter().map(UnnamedField::type_path).collect(),
+        TypeInfo::Tuple(info) => info.iter().map(UnnamedField::type_path).collect(),
+        TypeInfo::List(info) => vec![info.item_ty().type_path_table().path()],
+        TypeInfo::Array(info) => vec![info.item_ty().type_path_table().path()],
+        TypeInfo::Map(info) => vec![
+            info.key_ty().type_path_table().path(),
+            info.value_ty().type_path_table().path(),
+        ],
+        TypeInfo::Enum(info) => info
+            .iter()
+            .flat_map(|s| match s {
+                VariantInfo::Struct(s) => s.iter().map(NamedField::type_path).collect(),
+                VariantInfo::Tuple(s) => s.iter().map(UnnamedField::type_path).collect(),
+                VariantInfo::Unit(_) => vec![],
+            })
+            .collect(),
+        TypeInfo::Set(info) => vec![info.value_ty().type_path_table().path()],
+        TypeInfo::Opaque(_) => vec![],
+    };
 
-fn unit_export(id: u32, name: String) -> TypeExport {
-    TypeExport {
-        id,
-        name,
-        type_data: TypeData::Class(Class {
-            use_as: UseAs::all_supported(),
-            color: UNIT_COLOR.to_string(),
-            draw_fill: true,
-            members: vec![],
-        }),
-    }
-}
-
-#[allow(dead_code)]
-fn insert_value(a: &mut PropertyValue, b: PropertyValue) {
-    use PropertyValue as PV;
-    match (a, b) {
-        (PV::BoolValue(a), PV::BoolValue(b)) => {
-            *a = b;
-        }
-        (PV::FloatValue(a), PV::FloatValue(b)) => {
-            *a = b;
-        }
-        (PV::IntValue(a), PV::IntValue(b)) => {
-            *a = b;
-        }
-        (PV::ColorValue(a), PV::ColorValue(b)) => {
-            *a = b;
-        }
-        (PV::StringValue(a), PV::StringValue(b)) => {
-            *a = b;
-        }
-        (PV::FileValue(a), PV::FileValue(b)) => {
-            *a = b;
-        }
-        (PV::ObjectValue(a), PV::ObjectValue(b)) => {
-            *a = b;
-        }
-        (
-            PV::ClassValue {
-                property_type: property_type_a,
-                properties: properties_a,
-            },
-            PV::ClassValue {
-                property_type: property_type_b,
-                properties: properties_b,
-            },
-        ) => {
-            assert_eq!(property_type_a, &property_type_b);
-
-            for (name, value) in properties_b {
-                match properties_a.entry(name) {
-                    Entry::Occupied(mut a) => {
-                        insert_value(a.get_mut(), value);
-                    }
-                    Entry::Vacant(a) => {
-                        a.insert(value);
-                    }
-                }
-            }
-        }
-        _ => {
-            panic!("mismatched property values");
+    let mut all_deps = deps.clone();
+    for d in deps {
+        if let Some(t) = registry.get_with_type_path(d) {
+            let mut new_deps = dependencies(t, registry);
+            all_deps.append(&mut new_deps);
         }
     }
-}
-
-#[allow(dead_code)]
-fn insert_json_value(property: &mut PropertyValue, json: serde_json::Value) {
-    use PropertyValue as PV;
-    match (property, json) {
-        (PV::BoolValue(a), serde_json::Value::Bool(b)) => {
-            *a = b;
-        }
-        (PV::FloatValue(a), serde_json::Value::Number(b)) => {
-            *a = b.as_f64().unwrap() as f32;
-        }
-        (PV::IntValue(a), serde_json::Value::Number(b)) => {
-            *a = b.as_f64().unwrap() as i32;
-        }
-        (PV::ColorValue(a), serde_json::Value::String(b)) => {
-            *a = b.parse().unwrap();
-        }
-        (PV::StringValue(a), serde_json::Value::String(b)) => {
-            *a = b;
-        }
-        (PV::FileValue(a), serde_json::Value::String(b)) => {
-            *a = b;
-        }
-        (PV::ObjectValue(a), serde_json::Value::Number(b)) => {
-            *a = b.as_u64().unwrap() as u32;
-        }
-        (
-            PV::ClassValue {
-                property_type: _,
-                properties,
-            },
-            serde_json::Value::Object(b),
-        ) => {
-            for (name, value) in b {
-                if let Some(property) = properties.get_mut(&name) {
-                    insert_json_value(property, value);
-                }
-            }
-        }
-        (a, b) => {
-            panic!("mismatched property values: {:?} vs {:?}", a, b);
-        }
-    }
+    all_deps
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_insert_value() {
-        let mut a = PropertyValue::ClassValue {
-            property_type: "[f32; 3]".to_string(),
-            properties: tiled::Properties::from([
-                ("[0]".to_string(), PropertyValue::IntValue(0)),
-                ("[1]".to_string(), PropertyValue::IntValue(0)),
-                ("[2]".to_string(), PropertyValue::IntValue(0)),
-            ]),
-        };
-
-        let b = PropertyValue::ClassValue {
-            property_type: "[f32; 3]".to_string(),
-            properties: tiled::Properties::from([("[1]".to_string(), PropertyValue::IntValue(1))]),
-        };
-
-        let expected = PropertyValue::ClassValue {
-            property_type: "[f32; 3]".to_string(),
-            properties: tiled::Properties::from([
-                ("[0]".to_string(), PropertyValue::IntValue(0)),
-                ("[1]".to_string(), PropertyValue::IntValue(1)),
-                ("[2]".to_string(), PropertyValue::IntValue(0)),
-            ]),
-        };
-
-        insert_value(&mut a, b);
-        assert_eq!(a, expected);
-    }
-
-    #[test]
-    fn test_insert_json_value() {
-        let mut a = PropertyValue::ClassValue {
-            property_type: "[f32; 3]".to_string(),
-            properties: tiled::Properties::from([
-                ("[0]".to_string(), PropertyValue::IntValue(0)),
-                ("[1]".to_string(), PropertyValue::IntValue(0)),
-                ("[2]".to_string(), PropertyValue::IntValue(0)),
-            ]),
-        };
-
-        let b = serde_json::json!({
-            "[1]": 1,
-            "[2]": 2
-        });
-
-        let expected = PropertyValue::ClassValue {
-            property_type: "[f32; 3]".to_string(),
-            properties: tiled::Properties::from([
-                ("[0]".to_string(), PropertyValue::IntValue(0)),
-                ("[1]".to_string(), PropertyValue::IntValue(1)),
-                ("[2]".to_string(), PropertyValue::IntValue(2)),
-            ]),
-        };
-
-        insert_json_value(&mut a, b);
-        assert_eq!(a, expected);
-    }
+    use bevy::render::render_resource::encase::rts_array::Length;
 
     #[test]
     fn generate_with_entity() {
@@ -677,12 +699,13 @@ mod tests {
         registry.register::<ComponentA>();
 
         let exports = TypeExportRegistry::from_registry(&registry);
-        let export_type = exports.types.get(ComponentA::type_path()).unwrap();
-        assert_eq!(export_type.name, ComponentA::type_path().to_string());
+        let export_type = &exports.types.get(ComponentA::type_path()).unwrap();
+        assert_eq!(export_type.length(), 1);
+        assert_eq!(export_type[0].name, ComponentA::type_path().to_string());
         assert_eq!(
-            export_type.type_data,
+            export_type[0].type_data,
             TypeData::Class(Class {
-                use_as: UseAs::all_supported(),
+                use_as: USE_AS_PROPERTY.to_vec(),
                 color: DEFAULT_COLOR.to_string(),
                 draw_fill: true,
                 members: vec![Member {
@@ -705,12 +728,13 @@ mod tests {
         registry.register::<ComponentA>();
 
         let exports = TypeExportRegistry::from_registry(&registry);
-        let export_type = exports.types.get(ComponentA::type_path()).unwrap();
-        assert_eq!(export_type.name, ComponentA::type_path().to_string());
+        let export_type = &exports.types.get(ComponentA::type_path()).unwrap();
+        assert_eq!(export_type.length(), 1);
+        assert_eq!(export_type[0].name, ComponentA::type_path().to_string());
         assert_eq!(
-            export_type.type_data,
+            export_type[0].type_data,
             TypeData::Class(Class {
-                use_as: UseAs::all_supported(),
+                use_as: USE_AS_PROPERTY.to_vec(),
                 color: DEFAULT_COLOR.to_string(),
                 draw_fill: true,
                 members: vec![Member {
@@ -737,10 +761,11 @@ mod tests {
         registry.register::<EnumComponent>();
 
         let exports = TypeExportRegistry::from_registry(&registry);
-        let export_type = exports.types.get(EnumComponent::type_path()).unwrap();
-        assert_eq!(export_type.name, EnumComponent::type_path().to_string());
+        let export_type = &exports.types.get(EnumComponent::type_path()).unwrap();
+        assert_eq!(export_type.length(), 1);
+        assert_eq!(export_type[0].name, EnumComponent::type_path().to_string());
         assert_eq!(
-            export_type.type_data,
+            export_type[0].type_data,
             TypeData::Enum(Enum {
                 storage_type: StorageType::String,
                 values: vec!["VarA".to_string(), "VarB".to_string(), "VarC".to_string(),],
@@ -792,12 +817,16 @@ mod tests {
         registry.register::<StructComponent>();
 
         let exports = TypeExportRegistry::from_registry(&registry);
-        let export_type = exports.types.get(StructComponent::type_path()).unwrap();
-        assert_eq!(export_type.name, StructComponent::type_path().to_string());
+        let export_type = &exports.types.get(StructComponent::type_path()).unwrap();
+        assert_eq!(export_type.length(), 1);
         assert_eq!(
-            export_type.type_data,
+            export_type[0].name,
+            StructComponent::type_path().to_string()
+        );
+        assert_eq!(
+            export_type[0].type_data,
             TypeData::Class(Class {
-                use_as: UseAs::all_supported(),
+                use_as: USE_AS_PROPERTY.to_vec(),
                 color: DEFAULT_COLOR.to_string(),
                 draw_fill: true,
                 members: vec![
@@ -863,12 +892,13 @@ mod tests {
         registry.register::<TestVariant>();
 
         let exports = TypeExportRegistry::from_registry(&registry);
-        let export_type = exports.types.get(TestOuter::type_path()).unwrap();
-        assert_eq!(export_type.name, TestOuter::type_path().to_string());
+        let export_type = &exports.types.get(TestOuter::type_path()).unwrap();
+        assert_eq!(export_type.length(), 1);
+        assert_eq!(export_type[0].name, TestOuter::type_path().to_string());
         assert_eq!(
-            export_type.type_data,
+            export_type[0].type_data,
             TypeData::Class(Class {
-                use_as: UseAs::all_supported(),
+                use_as: USE_AS_PROPERTY.to_vec(),
                 color: DEFAULT_COLOR.to_string(),
                 draw_fill: true,
                 members: vec![Member {
@@ -880,6 +910,139 @@ mod tests {
                         "1": serde_json::Value::default(),
                     })
                 }]
+            })
+        );
+    }
+
+    #[test]
+    fn generate_non_simple_enum() {
+        #[derive(Reflect)]
+        struct TestStruct {
+            a: i32,
+            b: String,
+        }
+
+        #[derive(Component, Reflect)]
+        #[reflect(Component)]
+        enum EnumComponent {
+            VarA,
+            VarB(u32),
+            VarC(TestStruct),
+            VarD((u16, f32)),
+        }
+
+        let mut registry = TypeRegistry::new();
+        registry.register::<TestStruct>();
+        registry.register::<EnumComponent>();
+
+        let exports = TypeExportRegistry::from_registry(&registry);
+        let export_type = &exports.types.get(EnumComponent::type_path()).unwrap();
+
+        assert_eq!(export_type.length(), 5);
+        assert_eq!(
+            export_type[0].name,
+            EnumComponent::type_path().to_string() + ":::Variant"
+        );
+        assert_eq!(
+            export_type[0].type_data,
+            TypeData::Enum(Enum {
+                storage_type: StorageType::String,
+                values: vec![
+                    "VarA".to_string(),
+                    "VarB".to_string(),
+                    "VarC".to_string(),
+                    "VarD".to_string(),
+                ],
+                values_as_flags: false,
+            }),
+        );
+        assert_eq!(
+            export_type[1].name,
+            EnumComponent::type_path().to_string() + "::VarB"
+        );
+        assert_eq!(
+            export_type[1].type_data,
+            TypeData::Class(Class {
+                use_as: USE_AS_PROPERTY.to_vec(),
+                color: DEFAULT_COLOR.to_string(),
+                draw_fill: true,
+                members: vec![Member {
+                    name: "0".to_string(),
+                    property_type: None,
+                    type_field: FieldType::Int,
+                    value: serde_json::Value::default(),
+                }],
+            })
+        );
+        assert_eq!(
+            export_type[2].name,
+            EnumComponent::type_path().to_string() + "::VarC"
+        );
+        assert_eq!(
+            export_type[2].type_data,
+            TypeData::Class(Class {
+                use_as: USE_AS_PROPERTY.to_vec(),
+                color: DEFAULT_COLOR.to_string(),
+                draw_fill: true,
+                members: vec![Member {
+                    name: "0".to_string(),
+                    property_type: Some(TestStruct::type_path().to_string()),
+                    type_field: FieldType::Class,
+                    value: serde_json::Value::default(),
+                }],
+            })
+        );
+        assert_eq!(
+            export_type[3].name,
+            EnumComponent::type_path().to_string() + "::VarD"
+        );
+        assert_eq!(
+            export_type[3].type_data,
+            TypeData::Class(Class {
+                use_as: USE_AS_PROPERTY.to_vec(),
+                color: DEFAULT_COLOR.to_string(),
+                draw_fill: true,
+                members: vec![Member {
+                    name: "0".to_string(),
+                    property_type: Some("(u16, f32)".to_string()),
+                    type_field: FieldType::Class,
+                    value: serde_json::Value::default(),
+                }],
+            })
+        );
+        assert_eq!(export_type[4].name, EnumComponent::type_path().to_string());
+        assert_eq!(
+            export_type[4].type_data,
+            TypeData::Class(Class {
+                use_as: USE_AS_PROPERTY.to_vec(),
+                color: DEFAULT_COLOR.to_string(),
+                draw_fill: true,
+                members: vec![
+                    Member {
+                        name: ":variant".to_string(),
+                        property_type: Some(EnumComponent::type_path().to_string() + ":::Variant"),
+                        type_field: FieldType::Class,
+                        value: serde_json::json!("VarA"),
+                    },
+                    Member {
+                        name: "VarB".to_string(),
+                        property_type: Some(EnumComponent::type_path().to_string() + "::VarB"),
+                        type_field: FieldType::Class,
+                        value: serde_json::Value::default(),
+                    },
+                    Member {
+                        name: "VarC".to_string(),
+                        property_type: Some(EnumComponent::type_path().to_string() + "::VarC"),
+                        type_field: FieldType::Class,
+                        value: serde_json::Value::default(),
+                    },
+                    Member {
+                        name: "VarD".to_string(),
+                        property_type: Some(EnumComponent::type_path().to_string() + "::VarD"),
+                        type_field: FieldType::Class,
+                        value: serde_json::Value::default(),
+                    },
+                ],
             })
         );
     }