+++ src/properties/types_json.rs
@@ -77,18 +77,6 @@ pub(crate) enum UseAs {
     Project,
 }
 
-impl UseAs {
-    pub fn all_supported() -> Vec<UseAs> {
-        vec![
-            UseAs::Property,
-            UseAs::Map,
-            UseAs::Layer,
-            UseAs::Object,
-            UseAs::Tile,
-        ]
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;